requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
flask==3.0.0
flask-cors==4.0.0
scikit-learn>=1.3.0
nltk>=3.8
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
textblob>=0.17.0
wordcloud>=1.9.0
